requirements.txt
yt-dlp
demucs
openai-whisper
flask
pysubs2
